# perf: lessen memory consumption for emulated mget command