--- a/.github/workflows/test.yaml
+++ b/.github/workflows/test.yaml
@@ -246,6 +246,8 @@ jobs:
           - single
           - excessive_pipelining
           - pipelining_in_moderation
+          - original_mget
+          - emulated_mget
     env:
       REDIS_VERSION: '7.2'
       DOCKER_COMPOSE_FILE: 'compose.yaml'