--- a/test/prof_mem.rb
+++ b/test/prof_mem.rb
@@ -10,6 +10,8 @@ module ProfMem
   ATTEMPT_COUNT = 1000
   MAX_PIPELINE_SIZE = 100
   SLICED_NUMBERS = (1..ATTEMPT_COUNT).each_slice(MAX_PIPELINE_SIZE).freeze
+  ORIGINAL_MGET = (%w[MGET] + Array.new(40) { |i| "{key}#{i}" }).freeze
+  EMULATED_MGET = (%w[MGET] + Array.new(40) { |i| "key#{i}" }).freeze
   CLI_TYPES = %w[primary_only scale_read_random scale_read_latency pooled].freeze
   MODES = {
     single: lambda do |client_builder_method|
@@ -38,6 +40,14 @@ module ProfMem
           numbers.each { |i| pi.call('GET', i) }
         end
       end
+    end,
+    original_mget: lambda do |client_builder_method|
+      cli = send(client_builder_method)
+      ATTEMPT_COUNT.times { cli.call_v(ORIGINAL_MGET) }
+    end,
+    emulated_mget: lambda do |client_builder_method|
+      cli = send(client_builder_method)
+      ATTEMPT_COUNT.times { cli.call_v(EMULATED_MGET) }
     end
   }.freeze