* #356

## Before
```
allocated memory by gem
-----------------------------------
  14959968  redis-client-0.22.1
  14496373  redis-cluster-client/lib
    120832  other
     80439  socket
     11472  uri
```

## After
```
allocated memory by gem
-----------------------------------
  13359968  redis-client-0.22.1
   6496373  redis-cluster-client/lib
    120832  other
     63891  socket
     11472  uri
```

## But
```
ruby 3.2.2 (2023-03-30 revision e51014f9c0) [x86_64-linux]
Warming up --------------------------------------
      MGET: original   313.000 i/100ms
      MGET: emulated    54.000 i/100ms
Calculating -------------------------------------
      MGET: original      3.083k (±10.0%) i/s -     15.337k in   5.049041s
      MGET: emulated    521.688 (± 8.8%) i/s -      2.592k in   5.013249s

Comparison:
      MGET: original:     3083.2 i/s
      MGET: emulated:      521.7 i/s - 5.91x  slower
```